+++ ft_strmapi.c
@@ -6,7 +6,7 @@
 /*   By: susumuyagi <susumuyagi@student.42.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/23 12:14:37 by susumuyagi        #+#    #+#             */
-/*   Updated: 2023/05/23 13:17:09 by susumuyagi       ###   ########.fr       */
+/*   Updated: 2023/05/24 16:35:58 by susumuyagi       ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,9 +15,9 @@
 
 char	*ft_strmapi(char const *s, char (*f)(unsigned int, char))
 {
-	char	*ret;
-	size_t	len;
-	size_t	i;
+	char			*ret;
+	size_t			len;
+	unsigned int	i;
 
 	if (!s || !f)
 		return (NULL);
